script: Exact VulkanRegistry to be replaced by VulkanObject
